Implement VESA video setup
* Implement VESA video mode setup in the BIOS bootloader with
  mode enumeration and selection
* Add bootloader config file parser (loader.conf) for configuring
  resolution and other options
* Add ELF and symbol parsing libraries under libs/parse/
* Add VGA text mode support header

+++ kernel/arch/x86/lib/regs.h
@@ -11,7 +11,7 @@
 #define REAL_SEG(addr) (u16)(((int)(uintptr_t)(addr) & 0xffff0) >> 4)
 #define REAL_OFF(addr) (u16)((int)(uintptr_t)(addr) & 0x0000f)
 
-#define REAL_FLATTEN(seg, off) (((u32)(seg) << 4) + (u32)(off))
+#define REAL_FLATTEN(seg, off) (uintptr_t)(((u32)(seg) << 4) + (u32)(off))
 
 // 32 bit registers in the order of popping -- used for bios calls
 // https://faydoc.tripod.com/cpu/popa.htm
